removed old prefix command handlers
