+++ bot/handlers_test.go
